### 2019-02-26 #285 @Rekyt
Submission: rromeo
Submitting Author: Matthias Grenié (@Rekyt)
Repository: https://github.com/Rekyt/rromeo
Version submitted: 0.1.0
Version accepted: TBD

---

```
Package: rromeo
Title: An R Client for SHERPA/RoMEO API
Version: 0.1.0
  person("Matthias", "Grenié", email = REDACTED,
         role = c("aut", "cre"), comment = c(ORCID = "0000-0002-4659-7522")),
  person("Hugo", "Gruson",
         role = "aut", comment = c(ORCID = "0000-0002-4094-1476"))
  )
Description: Provides an R Client for SHERPA/RoMEO API
  <http://www.sherpa.ac.uk/romeo/apimanual.php> which indexes policies of
  journal regarding the archival of scientific manuscripts before and/or after
  peer-review as well as formatted manuscripts.
    httr,
    xml2
    vcr (>= 0.2.6),
    ISOcodes,
    mockery
```

This package is an R wrapper to the SHERPA/RoMEO API available online. It provides direct access to the API through R functions.

Specifically bibliometricians interested in retrieving data on archival policies of different journals. `rromeo` can be use to draw a quick portrait of the archival policies of a given scientific field, or to show differences among publisher practices.

To our knowledge there are no R packages providing access to this API.

https://github.com/ropensci/software-review/issues/273
