### 2019-02-27 #285 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @Rekyt  !

Here's the output from [goodpractice](https://github.com/MangoTheCat/goodpractice/). If you haven't used `goodpractice` it's an R package that checks a number of things with another package - most of which we agree with and want authors to follow. You don't need to address these now, it's info for reviewers to use to get started. The test coverage is already high, nice job.

```r
── GP rromeo ───────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 96% of code lines are covered by test
    cases.

    R/rromeo_publishers.R:199:NA
    R/rromeo_publishers.R:200:NA
    R/rromeo_publishers.R:202:NA
    R/rromeo_publishers.R:265:NA
    R/rromeo_publishers.R:267:NA
    ... and 3 more lines

  ✖ use '<-' for assignment instead of '='. '<-' is the standard, and R users and developers are used it and
    it is easier to read your code for them if you use '<-'.

    R/rromeo_base.R:3:13
    R/rromeo_base.R:7:13
    R/rromeo_base.R:20:16
    R/rromeo_base.R:21:12
    R/rromeo_journals.R:20:17
    ... and 119 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about
    80 characters wide. Try make your lines shorter than 80 characters

    R/utils.R:98:1
```

Seeking reviewers now 🕐

---

Could you please add the rOpenSci under review badge to your README?

Reviewers:
