### 2019-03-29 #285 @ottlngr
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

First of all, `rromeo` is a well-made package and very close to a positive review.

When reviewing a package, I always start with the vignette. The vignette that comes along with `rromeo` made a positive impression on me, as it summarizes the major usage of the package and offers detailed examples on the different query options as well as plotting examples. Sole point of criticism: Some of the outputs are too long and should be shortened for the rendered version of the vignette.

##### Packaging guidelines

Why I did not tick off `**Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines` yet:

- the output of `goodpractice`, especially the mash up of `<-` and `=`
- the missing or incorrect top-level documentation for `rromeo` which should be addressable by `?rromeo`. See http://r-pkgs.had.co.nz/man.html for a reference documentation.

##### Miscellaneous

- It's not an official guideline, but when implementing an API package, I always consider Hadley Wickham's [best practices](https://cran.r-project.org/web/packages/httr/vignettes/api-packages.html). One of them is to set a User Agent which refers to the use of your package. No necessity, but worth consideration.

- Regarding the API key: You provide the `key` argument in functions when needed, while not recommending to use that way to provide the API key. My opinion: If you do not recommend this way, do not offer it ;) And I would also not recommend it, and therefore I would recommend you to remove the `key` argument in all functions. Instead I would (my personal preference) reconstruct `rromeo` to always access the environment variable `SHERPAROMEO_KEY`. For abstraction and not forcing users to set the environment variable manually, I would introduce some kind of `rr_auth(key)` function, that stores the key in an environment variable. (#1)

##### Usage

I run the examples from the vignette and worked through the package this way. Here a few things I want to address explicitly:

- When using `jj_journal_issn()` with a malformed ISSN, I get the cryptic error `Error in FUN(X[[i]], ...) : ISSN is invalid, please check the format` and not the expected `Error in jj_journal_issn("foo") : ISSN is invalid, please check the format`. On the fly, I could not figure out why this happens.

- When using `rr_journal_name("Biostatistics", qtype = "contains")` I get a message stating that 17 journals match my query terms, that's fine. But I also get the warning `Select one journal from the provided list or enable multiple = TRUE` which, for me as a user, is incomprehensible. Because in my understanding, the actual warning is not that one should select one journal from the list, but that additional columns got dropped during the query due to performance.

- When using `rr_journal_name("Biostatistics", qtype = "contains", multiple = TRUE)` I get confusing messages, one saying that 17 journals match my query terms and the other saying that 2 journals match my query terms. But what is more relevant for me: By setting `multiple = TRUE` the behaviour of `rr_journal_name` changes completely. With `multiple = FALSE`, `rr_journal_name()` is actually some kind of a `rr_journal_find()` or `rr_journal_search()`. But with `multiple = TRUE` it is actually `rr_journal_issn()`. So my impression is, that the current family of `rr_journal_*()` functions is not consistent. I encourage the author to redefine the current `rr_journal_name()` with `multiple = FALSE` to a more generic `rr_journal_find()` or the like. This functions could be the entry point and provides ISSNs to `rr_journal_issn()` to get all the metadata for the journals. (#2)

- `r_publisher_all()` runs into an error: `Error in parse_generic(api_answer) : Object 'parsed' not found`

##### Conclusion

As soon as the author has improved the output of `goodpractice`, all formal requirements for a positive review are fulfilled.

Nevertheless, I hope that I could provide some helpful input on the package, especially from a user's sight. Regarding my quite specific "change requests" #1 and #2, I'm curious about the author's opinion.
