### 2019-03-30 #285 @brunaw
## Package Review

  - [**X**] As the reviewer I confirm that there are no [conflicts of interest](#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

#### Documentation

  - [**X**] **Installation instructions:** for the development version of package and any non-standard dependencies in README
  - [**X**] **Vignette(s)** demonstrating major functionality that runs successfully locally
  - [**X**] **Function Documentation:** for all exported functions in R help
  - [**X**] **Examples** for all exported functions in R Help that run successfully locally

#### Functionality

  - [**X**] **Installation:** Installation succeeds as documented.
  - [**X**] **Functionality:** Any functional claims of the software been confirmed.
  - [**X**] **Performance:** Any performance claims of the software been confirmed.
  - [**X**] **Automated tests:** Unit tests cover essential functions of the package
  - [**X**] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines

#### Final approval (post-review)

Estimated hours spent reviewing: 5

---

### Review Comments

- I haven't been able to find a statement of need or an indication fo the target audience of the package.

- For the `rr_journal_name()` function, it would be a good idea to not use the word `Warning` when printing the `multiple = TRUE` recommendation. A warning can be scary for new users as it looks like something went wrong.

- I find it a good practice to mention the existence of an API key before the examples, because this could be one of the first errors a user would encounter. Besides that, there aren't many explanations about how to set an APY key. You made recommendations, such as putting it in the `.Renviron` or `.Rprofile`, but it might not be obvious for the user how to do so. Some guidelines on this point can be helpful,  maybe write a vignette just to describe how to find, set and what are the best practices when using the API key. This will also help you avoid having an extra 'Details' section in most of the documentation files.

- The current identation in the `rr_journal_name()` documentation makes it look like you repeated the first two returns. A sub-item indentation after the second item might help make clear that the next following items are about the `multiple = TRUE` option.

- In the documentation of `rr_romeo_colour()` you point to the 'Details' section for the API key details, but the section it's not there. I reinforce the idea of having a separate vignette for the API key
 configuration.

- I got an error with `rr_publisher_all`:

> Error in rromeo:::parse_generic(api_answer) : object 'parsed' not found

I opened the source code in the function and found that my current configuration was:
> apicontrol
[1] "all"
> outcome
[1] "publisherFound"
> hits
[1] "3007"

This error was probably generated by the lack of an `else` statement in the function to define what happens when the previous conditionals were not satisfied. `parse_publisher()` worked for this configuration, but the way the conditionals are written doesn't allow it to return a result.

- I found the Contributor Code of Conduct, but that's different from a Contribution Guidelines. While the first specifies how a contributor should behave, the second provides instructions about the steps to
 contributing to the package. I would advise you to include it in the README or a CONTRIBUTING file, that could also contain the Code of Conduct.

- General comments addressing the following:

Q: Does the code comply with general principles in the Mozilla reviewing guide?
A: **Yes**.

Q: Does the package comply with the rOpenSci packaging guide?
A:  **Yes**.

Q: Are there improvements that could be made to the code style?
A: Reviewing the way the `if-else` statements are organized. One idea is using more `%in%` to avoid extensive `|` conditions and make sure there are conditions for all the possible cases.

Q: Is there code duplication in the package that should be reduced?
A: Only documentation duplication (the API key part).

Q: Are there user interface improvements that could be made?
A: No, the package looks very user-friendly and has plenty of useful messages.

Q: Are there performance improvements that could be made?
A: No.

Q: Is the documentation (installation instructions/vignettes/examples/demos) clear and sufficient? Does it use the principle of multiple points of entry i.e. takes into account the fact that any piece of documentation may be the first encounter the user has with the package and/or the tool/data it wraps?

A: **Yes**.

Q: Were functions and arguments named to work together to form a common, logical programming API that is easy to read, and autocomplete?

A: **Yes**.

### Conclusion

`rromeo` looks like a good package and will be even better with some small
improvements. I hope my comments are clear, so let me know if they aren't.

(thanks for the reminder, @Rekyt!)
