### 2019-04-11 #285 @Rekyt
We would like to thank @ottlngr and @brunaw for the time they took reviewing `rromeo` as well as their useful comments and suggestions.

As noted by both reviewers and the editor @sckott, `rromeo` didn't pass the checks from `goodpractice` it does now with Rekyt/rromeo@1ce7107 we've replaced all the `=` with `<-` as it is a more standard practice in the R community.

We'll first respond to @ottlngr comments:

> the missing or incorrect top-level documentation for rromeo which should be addressable by ?rromeo. See http://r-pkgs.had.co.nz/man.html for a reference documentation.

`rromeo` now has a good package-level documentation Rekyt/rromeo@184c955

> It's not an official guideline, but when implementing an API package, I always consider Hadley Wickham's best practices. One of them is to set a User Agent which refers to the use of your package. No necessity, but worth consideration.

We hadn't thought of adding a user-agent which is good practice and we thank @ottlngr for this great suggestion now implemented in Rekyt/rromeo@3b2afd5
The user-agent is now added as header to all queries.

> Regarding the API key: You provide the key argument in functions when needed, while not recommending to use that way to provide the API key. My opinion: If you do not recommend this way, do not offer it ;) And I would also not recommend it, and therefore I would recommend you to remove the key argument in all functions. Instead I would (my personal preference) reconstruct rromeo to always access the environment variable SHERPAROMEO_KEY. For abstraction and not forcing users to set the environment variable manually, I would introduce some kind of rr_auth(key) function, that stores the key in an environment variable. (#1)

For development reasons we still want to keep a `key` argument in all functions. However we now made an entire vignette dedicated to setting up the API key and the various means to achieve it. Rekyt/rromeo@967322a to Rekyt/rromeo@b45f218 see addition with https://github.com/Rekyt/rromeo/compare/58fb140..b45f218

We also now created an `rr_auth()` function which helps the user defining the API key. see Rekyt/rromeo@3c87b22

> When using jj_journal_issn() with a malformed ISSN, I get the cryptic error Error in FUN(X[[i]], ...) : ISSN is invalid, please check the format and not the expected Error in jj_journal_issn("foo") : ISSN is invalid, please check the format. On the fly, I could not figure out why this happens.

There was a missing `call. = FALSE` in the call to `stop` it is now fixed Rekyt/rromeo@b7b81c4

> When using rr_journal_name("Biostatistics", qtype = "contains") I get a message stating that 17 journals match my query terms, that's fine. But I also get the warning Select one journal from the provided list or enable multiple = TRUE which, for me as a user, is incomprehensible. Because in my understanding, the actual warning is not that one should select one journal from the list, but that additional columns got dropped during the query due to performance.
>
> When using rr_journal_name("Biostatistics", qtype = "contains", multiple = TRUE) I get confusing messages, one saying that 17 journals match my query terms and the other saying that 2 journals match my query terms. But what is more relevant for me: By setting multiple = TRUE the behaviour of rr_journal_name changes completely. With multiple = FALSE, rr_journal_name() is actually some kind of a rr_journal_find() or rr_journal_search(). But with multiple = TRUE it is actually rr_journal_issn(). So my impression is, that the current family of rr_journal_*() functions is not consistent. I encourage the author to redefine the current rr_journal_name() with multiple = FALSE to a more generic rr_journal_find() or the like. This functions could be the entry point and provides ISSNs to rr_journal_issn() to get all the metadata for the journals. (#2)

This was a very useful suggestion and the package now has two separate functions: `rr_journal_find()` which returns only the title and the ISSN of the journal and `rr_journal_name()` which always returns all the information on archiving policies of all journals. We thus removed the `multiple` argument. Rekyt/rromeo@b69df66 to Rekyt/rromeo@80a3aa8

Instead of a warning, there is now a message suggesting the usage of `rr_journal_name()` instead of `rr_journal_find()` when the user wants more information.

> r_publisher_all() runs into an error: Error in parse_generic(api_answer) : Object 'parsed' not found

There was an error in the `GET()` call and the function now works. Rekyt/rromeo@59a7b40

Regarding @brunaw comments:

> I haven't been able to find a statement of need or an indication fo the target audience of the package.

We now specify better the target audience of the package in the beginning of the README file. Rekyt/rromeo@c86dca8

> For the rr_journal_name() function, it would be a good idea to not use the word Warning when printing the multiple = TRUE recommendation. A warning can be scary for new users as it looks like something went wrong.

As explained above the function now gives a message instead of a warning.

> I find it a good practice to mention the existence of an API key before the examples, because this could be one of the first errors a user would encounter. Besides that, there aren't many explanations about how to set an APY key. You made recommendations, such as putting it in the .Renviron or .Rprofile, but it might not be obvious for the user how to do so. Some guidelines on this point can be helpful, maybe write a vignette just to describe how to find, set and what are the best practices when using the API key. This will also help you avoid having an extra 'Details' section in most of the documentation files.

Thanks for this very nice suggestion! We now include a full vignette regarding the way to set up an API key Rekyt/rromeo@967322a to Rekyt/rromeo@b45f218 see addition with https://github.com/Rekyt/rromeo/compare/58fb140..b45f218
We furthermore reordered the README to mention the API key at first. Rekyt/rromeo@75fbe2c

> The current identation in the rr_journal_name() documentation makes it look like you repeated the first two returns. A sub-item indentation after the second item might help make clear that the next following items are about the multiple = TRUE option.

The function is now split between `rr_journal_find()` (which returns only title and ISSN) and `rr_journal_name()` (which returns all the information) thus the documentation is now split between functions.

> In the documentation of rr_romeo_colour() you point to the 'Details' section for the API key details, but the section it's not there. I reinforce the idea of having a separate vignette for the API key configuration.

We now systematically point to the API vignette to setup the API key.

> I got an error with rr_publisher_all

Thank you for the detailed report of the error, it really helped to fix the bug. It is now fixed Rekyt/rromeo@59a7b40

> I found the Contributor Code of Conduct, but that's different from a Contribution Guidelines. While the first specifies how a contributor should behave, the second provides instructions about the steps to contributing to the package. I would advise you to include it in the README or a CONTRIBUTING file, that could also contain the Code of Conduct.

We now include a CONTRIBUTING file. Rekyt/rromeo@78fbe2c

> Q: Are there improvements that could be made to the code style?
> A: Reviewing the way the if-else statements are organized. One idea is using more %in% to avoid extensive | conditions and make sure there are conditions for all the possible cases.

Thank you for this thoughtful comment, it greatly improved the readability of our internal functions Rekyt/rromeo@0990b6f and Rekyt/rromeo@526c502

We hope we managed to answer and take into account all the comments and we would like to thank again both reviewers @brunaw and @ottlngr as well as the editor @sckott for their useful comments and suggestions!
