### 2019-04-15 #285 @sckott
Thanks for your detailed reply @Rekyt including links to commits, etc.

reviewers @brunaw  and @ottlngr : are you happy with the responses and changes made? Any further questions/suggestions?
