### 2019-04-25 #285 @ottlngr
Sorry for the late reply, @Rekyt @sckott

Thanks @Rekyt for the detailed comments on the reviews. I appreciate your changes on `rromeo` and therefore recommend to admit `rromeo` to rOpenSci :)
