### 2019-04-29 #285 @sckott
Approved!  Thanks again for your submission @Rekyt ! And thanks for your reviews @ottlngr and @brunaw  👌

To-dos:

* Please transfer the package  to `ropensci`- I've invited you to a team (you need to accept that invitation first, before transferring). You'll be made admin once you transfer
* add rOpenSci footer to README
`[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)`
* it's a good idea to add issue and PR template files to your .github folder, egs here https://github.com/ropensci/dotgithubfiles/

---

We've started putting together a bookdown with our best practice and tips, [this chapter](https://ropensci.github.io/dev_guide/collaboration.html) starts the 3d section that's about guidance for after onboarding. Please tell us what could be improved. The repo is at https://github.com/ropensci/dev_guide

Are you interested in doing a blog post for our blog https://ropensci.org/blog/ ?  either a short-form intro to it (https://ropensci.org/technotes/) or long-form post with more narrative about its development (https://ropensci.org/blog/). If so, we'll have our community manager @stefaniebutland get in touch with you on that
