### 2019-05-06 #285 @Rekyt
Hi, we would like to thank @brunaw and @ottlngr for checking the package again and recommending it!

@stefaniebutland, we would be interested in writing a regular blog post with the narrative about how we decided to build `rromeo` and Tues June 4 would be perfect for us!
We'll submit a PR with the draft of the post.

Many thanks!
