### 2019-05-27 #285 @stefaniebutland
@Rekyt @Bisaloo  Checking in. Is Tues June 4 still a good publication date for you? If so, please submit a draft May 28 or as soon as possible after that so we have time for any refinements
